test(gateway): fix journal fixtures, split custodian page tests, refresh docs map
