+++ ui/src/pages/custodian/custodian-page.test.ts
@@ -1,133 +1,10 @@
 /* @vitest-environment jsdom */
 
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import type {
-  GatewayBrowserClient,
-  GatewayEventFrame,
-  GatewayEventListener,
-} from "../../api/gateway.ts";
-import type {
-  ApplicationContext,
-  ApplicationGateway,
-  ApplicationGatewaySnapshot,
-} from "../../app/context.ts";
-import {
-  createApplicationContextProvider,
-  type ApplicationContextProvider,
-} from "../../test-helpers/application-context.ts";
+import type { GatewayBrowserClient } from "../../api/gateway.ts";
+import type { ApplicationGatewaySnapshot } from "../../app/context.ts";
 import { waitForFast } from "../../test-helpers/wait-for.ts";
-import "./custodian-page.ts";
-
-type TestCustodianPage = HTMLElement & {
-  onboarding: boolean;
-  updateComplete: Promise<boolean>;
-};
-
-type ContextHarness = {
-  context: ApplicationContext;
-  setGatewaySnapshot: (patch: Partial<ApplicationGatewaySnapshot>) => void;
-  setGatewayUrl: (gatewayUrl: string) => void;
-  setGatewayToken: (token: string) => void;
-  setGatewayBootstrapToken: (bootstrapToken: string) => void;
-  setGatewayDeviceToken: (deviceToken: string) => void;
-  emitGatewayEvent: (event: Pick<GatewayEventFrame, "event" | "payload">) => void;
-};
-
-function createContext(
-  request: ReturnType<typeof vi.fn>,
-  methods: string[] = ["openclaw.chat"],
-): ContextHarness {
-  const client = { request } as unknown as GatewayBrowserClient;
-  let snapshot: ApplicationGatewaySnapshot = {
-    client,
-    connected: true,
-    reconnecting: false,
-    hello: {
-      type: "hello-ok" as const,
-      protocol: 1,
-      auth: { role: "operator", scopes: ["operator.admin"] },
-      features: { methods },
-    },
-    assistantAgentId: "main",
-    sessionKey: "main",
-    lastError: null,
-    lastErrorCode: null,
-  };
-  const listeners = new Set<(snapshot: ApplicationGatewaySnapshot) => void>();
-  const eventListeners = new Set<GatewayEventListener>();
-  const connection = {
-    gatewayUrl: "ws://gateway.test/control",
-    token: "",
-    bootstrapToken: "",
-    password: "",
-  };
-  const gateway = {
-    get snapshot() {
-      return snapshot;
-    },
-    connection,
-    subscribe: (listener: (snapshot: ApplicationGatewaySnapshot) => void) => {
-      listeners.add(listener);
-      return () => listeners.delete(listener);
-    },
-    subscribeEvents: (listener: GatewayEventListener) => {
-      eventListeners.add(listener);
-      return () => eventListeners.delete(listener);
-    },
-  } as unknown as ApplicationGateway;
-  const context = {
-    gateway,
-    basePath: "",
-    navigate: vi.fn(),
-  } as unknown as ApplicationContext;
-  return {
-    context,
-    setGatewaySnapshot: (patch) => {
-      snapshot = { ...snapshot, ...patch };
-      for (const listener of listeners) {
-        listener(snapshot);
-      }
-    },
-    setGatewayUrl: (gatewayUrl) => {
-      connection.gatewayUrl = gatewayUrl;
-    },
-    setGatewayToken: (token: string) => {
-      connection.token = token;
-    },
-    setGatewayBootstrapToken: (value: string) => {
-      connection.bootstrapToken = value;
-    },
-    setGatewayDeviceToken: (deviceToken: string) => {
-      snapshot = {
-        ...snapshot,
-        hello: snapshot.hello
-          ? { ...snapshot.hello, auth: { ...snapshot.hello.auth, deviceToken } }
-          : snapshot.hello,
-      };
-    },
-    emitGatewayEvent: (event) => {
-      for (const listener of eventListeners) {
-        listener(event as GatewayEventFrame);
-      }
-    },
-  };
-}
-
-async function mountPage(
-  context: ApplicationContext,
-  options: { onboarding?: boolean } = {},
-): Promise<{
-  page: TestCustodianPage;
-  provider: ApplicationContextProvider;
-}> {
-  const provider = createApplicationContextProvider(context);
-  const page = document.createElement("openclaw-custodian-page") as TestCustodianPage;
-  page.onboarding = options.onboarding ?? true;
-  provider.append(page);
-  document.body.append(provider);
-  await page.updateComplete;
-  return { page, provider };
-}
+import { createContext, mountPage } from "./custodian-page.test-harness.ts";
 
 describe("custodian page", () => {
   beforeEach(() => {
@@ -137,142 +14,6 @@ describe("custodian page", () => {
   afterEach(() => {
     document.body.replaceChildren();
     vi.restoreAllMocks();
-  });
-
-  it("shows advertised recent changes and loads a short cursor page inline", async () => {
-    const request = vi.fn(async (method: string, params: Record<string, unknown>) => {
-      if (method === "openclaw.chat") {
-        return {
-          sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-          reply: "Hello.",
-          action: "none",
-        };
-      }
-      if (method === "openclaw.changes.list" && !params.beforeCursor) {
-        return {
-          entries: [
-            {
-              id: "system-agent-audit:3",
-              at: Date.now() - 5_000,
-              kind: "operation",
-              source: "system-agent",
-              summary: "Set config gateway.port",
-              changedPaths: ["gateway.port"],
-            },
-            {
-              id: "config-audit:2",
-              at: Date.now() - 10_000,
-              kind: "external-edit",
-              source: "external",
-              summary: "Configuration edited outside OpenClaw",
-              invalid: true,
-              opaqueChange: true,
-            },
-            {
-              id: "config-audit:0",
-              at: Date.now() - 15_000,
-              kind: "config-write",
-              source: "plugin-install",
-              summary: "Plugin installation updated configuration",
-            },
-          ],
-          nextCursor: "next-page",
-        };
-      }
-      if (method === "openclaw.changes.list" && params.beforeCursor === "next-page") {
-        return {
-          entries: [
-            {
-              id: "config-audit:1",
-              at: Date.now() - 20_000,
-              kind: "config-write",
-              source: "config-rpc",
-              summary: "Settings updated configuration: agents.defaults.model",
-            },
-          ],
-        };
-      }
-      throw new Error(`unexpected request ${method}`);
-    });
-    const harness = createContext(request, ["openclaw.chat", "openclaw.changes.list"]);
-    const { context } = harness;
-    const { page } = await mountPage(context, { onboarding: false });
-    await waitForFast(() =>
-      expect(request).toHaveBeenCalledWith("openclaw.chat", expect.anything(), expect.anything()),
-    );
-    await page.updateComplete;
-
-    page.querySelector<HTMLButtonElement>(".custodian__history-toggle")!.click();
-    await waitForFast(() =>
-      expect(request).toHaveBeenCalledWith("openclaw.changes.list", { limit: 50 }),
-    );
-    await page.updateComplete;
-
-    expect(page.querySelectorAll(".custodian__change-card")).toHaveLength(3);
-    // A cursor, not a full page, controls continuation so bounded journal scans
-    // can return fewer than 50 cards without hiding Load more.
-    expect(page.querySelector(".custodian__history-more")).not.toBeNull();
-    expect(page.querySelector(".custodian__change-source")?.textContent).toContain("system-agent");
-    expect(page.querySelector(".custodian__change-paths")?.hasAttribute("open")).toBe(false);
-    expect(page.querySelector(".custodian__change-card.is-invalid")?.textContent).toContain(
-      "did not pass configuration validation",
-    );
-    expect(page.querySelector(".custodian__change-card.is-invalid")?.textContent).toContain(
-      "Formatting or comments changed",
-    );
-    expect(
-      Array.from(page.querySelectorAll(".custodian__change-source")).map((node) =>
-        node.textContent?.trim(),
-      ),
-    ).toContain("plugin install");
-
-    page.querySelector<HTMLButtonElement>(".custodian__history-more")!.click();
-    await waitForFast(() =>
-      expect(request).toHaveBeenCalledWith("openclaw.changes.list", {
-        limit: 50,
-        beforeCursor: "next-page",
-      }),
-    );
-    await page.updateComplete;
-    expect(page.querySelectorAll(".custodian__change-card")).toHaveLength(4);
-    expect(page.querySelector(".custodian__history-more")).toBeNull();
-
-    page.querySelector<HTMLButtonElement>(".custodian__history-toggle")!.click();
-    await page.updateComplete;
-    page.querySelector<HTMLButtonElement>(".custodian__history-toggle")!.click();
-    await waitForFast(() =>
-      expect(
-        request.mock.calls.filter(
-          ([method, params]) => method === "openclaw.changes.list" && !params.beforeCursor,
-        ),
-      ).toHaveLength(2),
-    );
-    await page.updateComplete;
-    expect(page.querySelectorAll(".custodian__change-card")).toHaveLength(3);
-
-    harness.setGatewaySnapshot({
-      client: { request } as unknown as GatewayBrowserClient,
-    });
-    await waitForFast(() => expect(page.querySelector(".custodian__history")).toBeNull());
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__history")).toBeNull();
-    expect(page.querySelector(".custodian__history-toggle")?.getAttribute("aria-expanded")).toBe(
-      "false",
-    );
-  });
-
-  it("hides change history when the gateway does not advertise it", async () => {
-    const request = vi.fn().mockResolvedValue({
-      sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-      reply: "Hello.",
-      action: "none",
-    });
-    const { context } = createContext(request);
-    const { page } = await mountPage(context, { onboarding: false });
-    await waitForFast(() => expect(request).toHaveBeenCalledOnce());
-    await page.updateComplete;
-
-    expect(page.querySelector(".custodian__history-toggle")).toBeNull();
   });
 
   it("starts onboarding chat, renders typed choices, and sends the option reply", async () => {
@@ -754,405 +495,6 @@ describe("custodian page", () => {
     await waitForFast(() => expect(request).toHaveBeenCalledTimes(2));
     expect(request.mock.calls[1]?.[1]).not.toHaveProperty("welcomeVariant");
     expect(request.mock.calls[1]?.[1]).toMatchObject({ message: "status" });
-  });
-
-  it("shows a channel-error nudge but ignores routine events", async () => {
-    const request = vi.fn().mockResolvedValue({
-      sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-      reply: "Everything is healthy.",
-      action: "none",
-    });
-    const { context, emitGatewayEvent } = createContext(request);
-    const { page } = await mountPage(context, { onboarding: false });
-    await waitForFast(() => expect(request).toHaveBeenCalledOnce());
-
-    emitGatewayEvent({ event: "tick", payload: { ts: Date.now() } });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")).toBeNull();
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channels: {
-          telegram: {
-            enabled: false,
-            accounts: {
-              default: {
-                configured: true,
-                enabled: false,
-                running: true,
-                connected: false,
-              },
-            },
-          },
-        },
-      },
-    });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")).toBeNull();
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channelLabels: { telegram: "Telegram" },
-        channels: {
-          telegram: {
-            enabled: false,
-            accounts: {
-              default: { configured: true, enabled: false, connected: false },
-              work: { configured: true, enabled: true, running: true, connected: false },
-            },
-          },
-        },
-      },
-    });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")?.textContent).toContain(
-      "Telegram just disconnected",
-    );
-  });
-
-  it("does not report an intentionally stopped channel as disconnected", async () => {
-    const request = vi.fn().mockResolvedValue({
-      sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-      reply: "Everything is healthy.",
-      action: "none",
-    });
-    const { context, emitGatewayEvent } = createContext(request);
-    const { page } = await mountPage(context, { onboarding: false });
-    await waitForFast(() => expect(request).toHaveBeenCalledOnce());
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channels: {
-          telegram: {
-            configured: true,
-            enabled: true,
-            running: false,
-            connected: false,
-            healthState: "not-running",
-            restartPending: false,
-            reconnectAttempts: 0,
-            lastStopAt: 1_700_000_000_000,
-            lastError: "connection closed during the previous run",
-          },
-        },
-      },
-    });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")).toBeNull();
-  });
-
-  it("does not report a recovered channel with a retained error", async () => {
-    const request = vi.fn().mockResolvedValue({
-      sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-      reply: "Everything is healthy.",
-      action: "none",
-    });
-    const { context, emitGatewayEvent } = createContext(request);
-    const { page } = await mountPage(context, { onboarding: false });
-    await waitForFast(() => expect(request).toHaveBeenCalledOnce());
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channels: {
-          telegram: {
-            configured: true,
-            enabled: true,
-            running: true,
-            healthState: "healthy",
-            lastError: "connection closed during the previous run",
-          },
-        },
-      },
-    });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")).toBeNull();
-  });
-
-  it("reports a channel that fails before its first start", async () => {
-    const request = vi.fn().mockResolvedValue({
-      sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-      reply: "Everything is healthy.",
-      action: "none",
-    });
-    const { context, emitGatewayEvent } = createContext(request);
-    const { page } = await mountPage(context, { onboarding: false });
-    await waitForFast(() => expect(request).toHaveBeenCalledOnce());
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channelLabels: { telegram: "Telegram" },
-        channels: {
-          telegram: {
-            configured: true,
-            enabled: true,
-            running: false,
-            connected: false,
-            restartPending: false,
-            reconnectAttempts: 0,
-            healthState: "not-running",
-            lastError: "failed to initialize transport",
-          },
-        },
-      },
-    });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")?.textContent).toContain("Telegram is degraded");
-  });
-
-  it("reports a failed restart after an earlier clean stop", async () => {
-    const request = vi.fn().mockResolvedValue({
-      sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-      reply: "Everything is healthy.",
-      action: "none",
-    });
-    const { context, emitGatewayEvent } = createContext(request);
-    const { page } = await mountPage(context, { onboarding: false });
-    await waitForFast(() => expect(request).toHaveBeenCalledOnce());
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channelLabels: { telegram: "Telegram" },
-        channels: {
-          telegram: {
-            configured: true,
-            enabled: true,
-            running: false,
-            restartPending: false,
-            reconnectAttempts: 0,
-            healthState: "not-running",
-            lastStopAt: 1_700_000_000_000,
-            lastStartAt: 1_700_000_001_000,
-            lastError: "failed to initialize transport",
-          },
-        },
-      },
-    });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")?.textContent).toContain("Telegram is degraded");
-  });
-
-  it("reports a current failed probe for an intentionally stopped channel", async () => {
-    const request = vi.fn().mockResolvedValue({
-      sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-      reply: "Everything is healthy.",
-      action: "none",
-    });
-    const { context, emitGatewayEvent } = createContext(request);
-    const { page } = await mountPage(context, { onboarding: false });
-    await waitForFast(() => expect(request).toHaveBeenCalledOnce());
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channelLabels: { telegram: "Telegram" },
-        channels: {
-          telegram: {
-            configured: true,
-            enabled: true,
-            running: false,
-            restartPending: false,
-            reconnectAttempts: 0,
-            healthState: "not-running",
-            lastStopAt: 1_700_000_001_000,
-            lastStartAt: 1_700_000_000_000,
-            probe: { ok: false },
-          },
-        },
-      },
-    });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")?.textContent).toContain("Telegram is degraded");
-  });
-
-  it("shows a channel disconnect from the aggregate health row", async () => {
-    const request = vi.fn().mockResolvedValue({
-      sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-      reply: "Everything is healthy.",
-      action: "none",
-    });
-    const { context, emitGatewayEvent } = createContext(request);
-    const { page } = await mountPage(context, { onboarding: false });
-    await waitForFast(() => expect(request).toHaveBeenCalledOnce());
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channelLabels: { telegram: "Telegram" },
-        channels: {
-          telegram: { configured: true, running: true, connected: false },
-        },
-      },
-    });
-    await page.updateComplete;
-
-    expect(page.querySelector(".custodian__nudge")?.textContent).toContain(
-      "Telegram just disconnected",
-    );
-  });
-
-  it("clears a pending event nudge when the gateway identity changes", async () => {
-    const request = vi.fn().mockResolvedValue({
-      sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-      reply: "Everything is healthy.",
-      action: "none",
-    });
-    const { context, emitGatewayEvent, setGatewaySnapshot, setGatewayUrl } = createContext(request);
-    const { page } = await mountPage(context, { onboarding: false });
-    await waitForFast(() => expect(request).toHaveBeenCalledOnce());
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channels: { telegram: { configured: true, running: true, connected: false } },
-      },
-    });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")).not.toBeNull();
-
-    setGatewayUrl("ws://gateway-b.test/control");
-    setGatewaySnapshot({
-      client: { request } as unknown as GatewayBrowserClient,
-      connected: true,
-      reconnecting: false,
-    });
-    await waitForFast(() => expect(request).toHaveBeenCalledTimes(2));
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")).toBeNull();
-
-    emitGatewayEvent({
-      event: "health",
-      payload: { configReload: { hotReloadStatus: "disabled" }, channels: {} },
-    });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")?.textContent).toContain(
-      "Configuration reload stopped",
-    );
-  });
-
-  it("dismisses event nudges for the rest of the page visit", async () => {
-    const request = vi.fn().mockResolvedValue({
-      sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-      reply: "Everything is healthy.",
-      action: "none",
-    });
-    const { context, emitGatewayEvent } = createContext(request);
-    const { page } = await mountPage(context, { onboarding: false });
-    await waitForFast(() => expect(request).toHaveBeenCalledOnce());
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channels: { telegram: { configured: true, running: true, connected: false } },
-      },
-    });
-    await page.updateComplete;
-    page.querySelector<HTMLButtonElement>(".custodian__nudge-dismiss")!.click();
-    await page.updateComplete;
-
-    emitGatewayEvent({
-      event: "health",
-      payload: { configReload: { hotReloadStatus: "disabled" }, channels: {} },
-    });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")).toBeNull();
-    expect(request).toHaveBeenCalledOnce();
-  });
-
-  it("replaces a pending nudge only with a more severe event", async () => {
-    const request = vi.fn().mockResolvedValue({
-      sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-      reply: "Everything is healthy.",
-      action: "none",
-    });
-    const { context, emitGatewayEvent } = createContext(request);
-    const { page } = await mountPage(context, { onboarding: false });
-    await waitForFast(() => expect(request).toHaveBeenCalledOnce());
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channelLabels: { telegram: "Telegram" },
-        channels: { telegram: { configured: true, healthState: "stale-socket" } },
-      },
-    });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")?.textContent).toContain("Telegram is degraded");
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channelLabels: { discord: "Discord" },
-        channels: { discord: { configured: true, healthState: "stale-socket" } },
-      },
-    });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")?.textContent).toContain("Telegram is degraded");
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channelLabels: { discord: "Discord" },
-        channels: { discord: { configured: true, running: true, connected: false } },
-      },
-    });
-    await page.updateComplete;
-    expect(page.querySelector(".custodian__nudge")?.textContent).toContain(
-      "Discord just disconnected",
-    );
-  });
-
-  it("sends a real message when an event nudge is clicked", async () => {
-    const request = vi.fn().mockResolvedValue({
-      sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-      reply: "Everything is healthy.",
-      action: "none",
-    });
-    const { context, emitGatewayEvent } = createContext(request);
-    const { page } = await mountPage(context, { onboarding: false });
-    await waitForFast(() => expect(request).toHaveBeenCalledOnce());
-
-    emitGatewayEvent({
-      event: "health",
-      payload: {
-        channels: {
-          telegram: { configured: true, tokenStatus: "configured_unavailable" },
-        },
-      },
-    });
-    await page.updateComplete;
-    page.querySelector<HTMLButtonElement>(".custodian__nudge-action")!.click();
-
-    await waitForFast(() => expect(request).toHaveBeenCalledTimes(2));
-    expect(request.mock.calls[1]?.[1]).toMatchObject({
-      message: "what happened with telegram authentication?",
-    });
-    expect(page.textContent).toContain("what happened with telegram authentication?");
-    expect(page.querySelector(".custodian__nudge")).toBeNull();
-  });
-
-  it("never shows event nudges during onboarding", async () => {
-    const request = vi.fn().mockResolvedValue({
-      sessionId: "control-ui-onboarding-00000000-0000-4000-8000-000000000001",
-      reply: "Welcome.",
-      action: "none",
-    });
-    const { context, emitGatewayEvent } = createContext(request);
-    const { page } = await mountPage(context, { onboarding: true });
-    await waitForFast(() => expect(request).toHaveBeenCalledOnce());
-
-    emitGatewayEvent({
-      event: "health",
-      payload: { configReload: { hotReloadStatus: "disabled" }, channels: {} },
-    });
-    await page.updateComplete;
-
-    expect(page.querySelector(".custodian__nudge")).toBeNull();
   });
 
   it("starts a fresh welcome when onboarding mode changes", async () => {
